Unify bid and ask hashmaps
Experiment to try to increase performance by using one hashmap.
Results are mixed, and not sure if there is much gain at the existing
benchmarks. But it is for sure increases unpredictability of the benchmarks,
therefore likely the whole system, and that is not cosidered a positive quality.
The tradeoff does not worth it.